Move python code to a module
